requirements.txt
asgiref==3.9.2
Django==5.2.6
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
